perfect assignment

+++ server/app.py
@@ -1,29 +1,3 @@
-# from flask import Flask, request, make_response, jsonify
-# from flask_cors import CORS
-# from flask_migrate import Migrate
-
-# from models import db, Message
-
-# app = Flask(__name__)
-# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///app.db'
-# app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-# app.json.compact = False
-
-# CORS(app)
-# migrate = Migrate(app, db)
-
-# db.init_app(app)
-
-# @app.route('/messages')
-# def messages():
-#     return ''
-
-# @app.route('/messages/<int:id>')
-# def messages_by_id(id):
-#     return ''
-
-# if __name__ == '__main__':
-#     app.run(port=5555)
 from flask import Flask, request, make_response, jsonify
 from flask_cors import CORS
 from flask_migrate import Migrate
@@ -61,7 +35,10 @@ def create_message():
 
 @app.route('/messages/<int:id>', methods=['PATCH'])
 def update_message(id):
-    message = Message.query.get_or_404(id)
+    message = db.session.get(Message, id)
+    if message is None:
+        return make_response(jsonify({"error": "Message not found"}), 404)
+
     data = request.get_json()
     body = data.get('body')
 
@@ -74,7 +51,10 @@ def update_message(id):
 
 @app.route('/messages/<int:id>', methods=['DELETE'])
 def delete_message(id):
-    message = Message.query.get_or_404(id)
+    message = db.session.get(Message, id)
+    if message is None:
+        return make_response(jsonify({"error": "Message not found"}), 404)
+    
     db.session.delete(message)
     db.session.commit()
     return make_response(jsonify({"message": "Message deleted successfully"}), 204)
